plugins/writing-workflow/requirements.txt
# Writing Workflow Plugin Requirements

# Core dependencies
anthropic>=0.40.0
requests>=2.31.0
python-dateutil>=2.8.2

# Content quality tools
textstat>=0.7.3
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Visual content generation
Pillow>=10.0.0
playwright>=1.40.0

# PowerPoint generation
python-pptx>=0.6.21

# Optional: Google Docs publishing
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
